requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
chromadb==0.4.22
openai==1.10.0
langchain==0.1.4
langchain-openai==0.0.5
pypdf2==3.0.1
python-docx==1.1.0
sentence-transformers==2.2.2
python-dotenv==1.0.0
aiofiles==23.2.1
tiktoken==0.5.2
